Restructure contribution code and separate providers
