+++ package.json
@@ -20,7 +20,6 @@
     "@types/jest": "^26.0.20",
     "@types/node": "^14.14.28",
     "@types/node-fetch": "^2.5.8",
-    "@types/psl": "^1.1.0",
     "@types/uuid": "^8.3.0",
     "@types/ws": "^7.4.0",
     "@typescript-eslint/eslint-plugin": "^4.15.1",
@@ -38,7 +37,6 @@
     "elliptic": "^6.5.4",
     "express": "^4.17.1",
     "node-fetch": "^2.6.1",
-    "psl": "^1.8.0",
     "uuid": "^8.3.2",
     "ws": "^7.4.3"
   }
